+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.6.6] - 2025-11-09
+
+### Beta
+- Version 1.6.6 release
+
+### Technical Details
+- Version: 1.6.6
+- Release Type: beta
+- Build Number: 202511092132
+- Release Date: 2025-11-09T21:32:54.216Z
+
+
 ## [1.6.5] - 2025-11-08
 
 ### Beta